# [feature] Implement functional filter capabilities for aggregate hashmaps