It turns out that using a functional approach was feasible after all (since after a bit of re-shuffling the code, the overhead for a default non-filter scenario is negligible, and even a filter call doesn't introduce enough overhead to significantly slow down operations on the `AggFlowMap` - after all we're operating on aggregate data already, the core hashmap or even the iterator code isn't touched).
I don't think this rather "edge"-y case and the small penalty warrants over-optimizing as long as the default case complexity is constant. If we need to squeeze out the last few percent we can do so in a future iteration.
```
                                     │    sec/op    │   sec/op     vs base               │
HashMapIterator-4                       982.8µ ± 2%   990.4µ ± 3%       ~ (p=0.436 n=10)
HashMapMetaIterator-4                   1.178m ± 3%   1.217m ± 2%  +3.34% (p=0.000 n=10)
HashMapMetaIteratorWithFilter/miss-4                  1.388m ± 2%
HashMapMetaIteratorWithFilter/hit-4                   1.559m ± 0%
```

@els0r PTAL